voice-server/requirements.txt
flask>=3.0
openai-whisper>=20231117
soundfile>=0.12
numpy>=1.24
openai>=1.0
